Cleanup, and a couple of natives changed
Changed local playerPed to local ped, insted of using to different variable names for PlayerPedId()

Changed GetPlayerPed(-1) in to PlayerPedId() and using the local ped var insted, to follow the code schema i used everywhere.

These updates are not testet and wont be made in to a release before it's testet.